# ccd-js-gen: Improve performance and CLI progress reporting