## Purpose

Improve performance and information printed for the CLI.
Generating code for wCCD went from `~3958ms` to `~2670ms` (on my machine) and only generating the typescript files is done in `~324ms`, see the example CLI output below for more details on where the time is spent.

Should be merged after #324.

<details>
  <summary>Example CLI output</summary>

  ```
Generating smart contract clients...

[1/20] 18ms    Parse smart contract module.

[2/20] 28ms    Generate base statements in module.

[3/20] 8ms     Generate initialize statements for 'cis2_wCCD' in module.

[4/20] 67ms    Generate base statements for 'cis2_wCCD'.

[5/20] 8ms     Generate statements for 'cis2_wCCD.wrap'.

[6/20] 9ms     Generate statements for 'cis2_wCCD.unwrap'.

[7/20] 13ms    Generate statements for 'cis2_wCCD.updateAdmin'.

[8/20] 9ms     Generate statements for 'cis2_wCCD.setPaused'.

[9/20] 10ms    Generate statements for 'cis2_wCCD.setMetadataUrl'.

[10/20] 11ms   Generate statements for 'cis2_wCCD.transfer'.

[11/20] 12ms   Generate statements for 'cis2_wCCD.updateOperator'.

[12/20] 10ms   Generate statements for 'cis2_wCCD.balanceOf'.

[13/20] 15ms   Generate statements for 'cis2_wCCD.operatorOf'.

[14/20] 15ms   Generate statements for 'cis2_wCCD.tokenMetadata'.

[15/20] 13ms   Generate statements for 'cis2_wCCD.view'.

[16/20] 18ms   Generate statements for 'cis2_wCCD.supports'.

[17/20] 18ms   Generate statements for 'cis2_wCCD.setImplementors'.

[18/20] 14ms   Generate statements for 'cis2_wCCD.upgrade'.

[19/20] 6ms    Saving generated TypeScript files.

[20/20] 2439ms Emitting JavaScript files.
```

</details>


## Changes

_No change in the generated smart contract client code from this PR._

- Build up structures instead of manipulating source files during codegen as [recommended by ts-morph](https://ts-morph.com/manipulation/performance#performance-tip-work-with-structures-instead).
- Add more detailed logging and timing for printing progress when using the CLI.

## Checklist

- [x] My code follows the style of this project.
- [x] The code compiles without warnings.
- [x] I have performed a self-review of the changes.
- [x] I have documented my code, in particular the intent of the
      hard-to-understand areas.
- [x] (If necessary) I have updated the CHANGELOG.